third_party/blink/web_tests/http/tests/devtools/runtime/runtime-es6-setSymbolPropertyValue-expected.txt

Tests editing Symbol properties.


Running: testSetUp
runtime-es6-setSymbolPropertyValue.js:21 ===== Initial =====
runtime-es6-setSymbolPropertyValue.js:22 2
runtime-es6-setSymbolPropertyValue.js:23 

Running: testSetSymbolPropertyValue
runtime-es6-setSymbolPropertyValue.js:21 ===== Set property =====
runtime-es6-setSymbolPropertyValue.js:22 3
runtime-es6-setSymbolPropertyValue.js:23 

Running: testDeleteSymbolProperty
runtime-es6-setSymbolPropertyValue.js:21 ===== Delete property =====
runtime-es6-setSymbolPropertyValue.js:22 undefined
runtime-es6-setSymbolPropertyValue.js:23 
